Rename settings route file
