+++ test/routes/settings.test.tsx
@@ -4,7 +4,7 @@ import { fromPartial } from '@total-typescript/shoehorn';
 import type { ActionFunctionArgs } from 'react-router';
 import { createRoutesStub } from 'react-router';
 import type { AuthHelper } from '../../app/auth/auth-helper.server';
-import SettingsComp, { action as settingsAction, loader } from '../../app/routes/settings.$';
+import SettingsComp, { action as settingsAction, loader } from '../../app/routes/settings';
 import type { SettingsService } from '../../app/settings/SettingsService.server';
 
 describe('settings', () => {
